backend/make_itinerary_api/src/requirements.txt
openai
functions_framework
googlemaps
firebase_admin
slackclient
python-dotenv
openmeteo-requests
requests-cache
retry-requests
numpy
pandas